FIX: all convois over 65 tons were too light
git-svn-id: svn://tron.homeunix.org/simutrans/simutrans/trunk@5825 8aca7d54-2c30-db11-9de9-000461428c89

+++ besch/vehikel_besch.h
@@ -69,7 +69,7 @@ private:
 	uint16  zuladung;
 	uint16	loading_time;	// time per full loading/unloading
 	uint16  geschw;
-	uint16  gewicht;
+	uint32  gewicht;
 	uint16  axle_load;
 	uint32  leistung;
 	uint16  running_cost;
